% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_max_snps.R
\name{find_max_snps}
\title{Find SNPs with Highest LOD}
\usage{
find_max_snps(snp_pos, snp_id, snp_lod, max_peak = NULL, haplo_peak_drop = NULL)
}

\description{
Internal function for \code{\link[=plot_scan2]{plot_scan2()}} and \code{\link[=plot_genes2]{plot_genes2()}} to find SNPs with highest LOD score within genomic interval.
}


\arguments{
\item{snp_pos}{Position of SNPs obtained from scan1info of \code{\link[=scan1snpasso]{scan1snpasso()}}.}
\item{snp_id}{SNP IDs obtained from scan1info of \code{\link[=scan1snpasso]{scan1snpasso()}}.}
\item{snp_lod}{SNP LOD scores obtained from scan1output of \code{\link[=scan1snpasso]{scan1snpasso()}}.}
\item{max_peak}{Numeric value in Mbps of the location of the haplotype's QTL peak. This can be obtained from \code{\link[=max_scan1]{max_scan1()}} 'pos' column.}
\item{haplo_peak_drop}{A vector of length two containing the range of Mbps that have LOD scores within the QTL peak drop. This vector should be the output from \code{\link[=lod_drop_rng]{lod_drop_rng()}}.}

}
