% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpdf_devoff.R
\name{jpdf_devoff}
\title{Shut down of pdf device}
\usage{
jpeg_devoff(pdfname, mypattern, ...)}

\description{
Shuts down the device from \code{\link[=jpeg_pdf]{jpef_pdf()}}.
}


\arguments{
\item{pdfname}{Name of the pdf file used to save jpeg images.}
\item{mypattern}{'mypattern' from output list returned by \code{\link[=jpeg_pdf]{jpef_pdf()}}.}
\item{...}{Additional parameters passed to image_write function from the Magick package.}
}
