% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpeg_pdf.R
\name{jpeg_pdf}
\title{Allows jpeg images to be converted to pdf}
\usage{
jpeg_pdf(pdfname, mypattern = "MYTEMPJPEG", ...)}

\description{
Conversion of jpeg image(s) to pdf format.
}


\arguments{
\item{pdfname}{Name of the pdf file to save jpeg images.}
\item{mypattern}{Pattern name of jpeg files. This should not be changed.}
\item{...}{Additional parameters passed to the convert function from ImageMagick.}
}
