% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lod_drop_rng.R
\name{lod_drop_rng}
\title{Find Haplotype Region with LOD scores within QTL Peak Drop}
\usage{
lod_drop_rng(markers, scan1_output, max_chr, max_peak, hilit_drop)
}

\description{
This function finds the haplotype region with LOD scores within the QTL peak drop.
}


\arguments{
\item{markers}{A dataframe of marker positions.}
\item{scan1_output}{Output of \code{\link[=scan1]{scan1()}}.}
\item{max_chr}{Character value of the haplotype's QTL peak chromosome. This can be obtained from \code{\link[=max_scan1]{max_scan1()}} 'chr' column.}
\item{max_peak}{Numeric value of the haplotype's QTL peak LOD score. This can be obtained from \code{\link[=max_scan1]{max_scan1()}} third column.}
\item{hilit_drop}{Positions with LOD scores within this amount of the QTL peak will be returned.}
}
