% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_int.R
\name{plot_int}
\alias{plot_int}
\title{Plot the additive, interactive, and difference genome scan}
\usage{
plot_int(add_scan, int_scan, diff_scan, map, lodcolumn = 1, chr = NULL)
}

\arguments{
\item{add_scan}{Output of \code{\link[=scan1]{scan1()}} without interactive covariates.}
\item{int_scan}{Output of \code{\link[=scan1]{scan1()}} with interactive covariates.}
\item{diff_scan}{Scan1 matrix containing the difference in LOD scores between the int_scan and add_scan.}

\item{map}{A list of vectors of marker positions, as produced by
\code{\link[=insert_pseudomarkers]{insert_pseudomarkers()}}.}

\item{lodcolumn}{LOD score column to plot (a numeric index, or a
character string for a column name). Only one value allowed.}

\item{chr}{Selected chromosomes to plot; a vector of character
strings.}
}



\description{
Utilizes \code{\link[=plot_scan1]{plot_scan1()}} to plot three LOD curves (additive, interactive, and difference) for a genome scan.
}
\section{Hidden graphics parameters}{

No hidden graphics parameters.
}
