% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_int_coeffCC.R
\name{plot_int_coeffCC}
\alias{plot_int_coeffCC}
\title{Plot sex-specific allele effects}
\usage{
plot_int_coeffCC(f_blup, m_blup, map, int_scan, add_scan, diff_scan, lodcolumn, chr)
}
\arguments{
\item{f_blup}{}

\item{m_blup}{}

\item{map}{}

\item{int_scan}{}

\item{add_scan}{}

\item{diff_scan}{}

\item{lodcolumn}{}

\item{chr}{}

}
\description{
Fixing this function...
}
