% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scan2.R
\name{plot_scan2}
\alias{plot_scan2}
\title{Plot a genome scan}
\usage{
plot_scan2(x, map, snp_lod, snpinfo, max_peak,
    haplo_peak_drop = NULL, lodcolumn=1, chr=NULL, add=FALSE, gap=25, ...)

\method{plot}{scan1}(x, map, lodcolumn = 1, chr = NULL, add = FALSE,
  gap = 25, ...)
}

\description{
Modification of qtl2's \code{\link[=plot_scan1]{plot_scan1()}} function to fit \code{\link[=plot_snpasso2]{plot_snpasso2()}} usage.
}

\arguments{
\item{x}{Output of \code{\link[=scan1]{scan1()}}.}

\item{map}{A list of vectors of marker positions, as produced by
\code{\link[=insert_pseudomarkers]{insert_pseudomarkers()}}.}


\item{snp_lod}{The scan1output of \code{\link[=scan1snpasso]{scan1snapsso()}}.}


\item{snpinfo}{Data frame with SNP information with the following
columns (the last three are generally derived from with
\code{\link[=index_snps]{index_snps()}}):
\itemize{
\item \code{chr} - Character string or factor with chromosome
\item \code{pos} - Position (in same units as in the \code{"map"}
attribute in \code{genoprobs}.
\item \code{sdp} - Strain distribution pattern: an integer, between
1 and \eqn{2^n - 2} where \eqn{n} is the number of strains, whose
binary encoding indicates the founder genotypes
\item \code{snp} - Character string with SNP identifier (if
missing, the rownames are used).
\item \code{index} - Indices that indicate equivalent
groups of SNPs.
\item \code{intervals} - Indexes that indicate which marker
intervals the SNPs reside.
\item \code{on_map} - Indicate whether SNP coincides with a marker
in the \code{genoprobs}
}}

\item{max_peak}{Numeric value in Mbps of the location of the haplotype's QTL peak. This can be obtained from \code{\link[=max_scan1]{max_scan1()}} 'pos' column.}

\item{haplo_peak_drop}{A vector of length two containing the range of Mbps that have LOD scores within the QTL peak drop. This vector should be the output from \code{\link[=lod_drop_rng]{lod_drop_rng()}}.}

\item{lodcolumn}{LOD score column to plot (a numeric index, or a
character string for a column name). Only one value allowed.}

\item{chr}{Selected chromosomes to plot; a vector of character
strings.}

\item{add}{If TRUE, add to current plot (must have same map and
chromosomes).}

\item{gap}{Gap between chromosomes.}

\item{...}{Additional graphics parameters.}
}


\section{Hidden graphics parameters}{

A number of graphics parameters can be passed via \code{...}. For
example, \code{bgcolor} to control the background color and
\code{altbgcolor} to control the background color on alternate chromosomes.
\code{col} controls the color of lines/curves; \code{altcol} can be used if
you want alternative chromosomes in different colors.
These are not included as formal parameters in order to avoid
cluttering the function definition.
}

