% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snpasso2.R
\name{plot_snpasso2}
\alias{plot_snpasso2}
\title{Plot SNP association}
\usage{
plot_snpasso2(scan1output, snpinfo, max_peak, haplo_peak_drop = NULL,
  haplo_scan = NULL, genes = NULL, lodcolumn = 1,
  show_all_snps = TRUE, chr = NULL, add = FALSE, drop_hilit = NA,
  col_hilit = "violetred", col = "darkslateblue", gap = 25,
  minlod = 0, ...)
}
\description{
Modified version of qtl2's \code{\link[=plot_snpasso]{plot_snpasso()}} function to allow plotting of the haplotype's QTL peak (assuming the region is set around the peak) and to easily identify which gene the SNP with the highest LOD score overlap. A solid black vertical line represents the haplotype's QTL peak, while a solid violet-red line represents the SNP with the highest LOD score. An optional parameter, haplo_peak_drop, will plot dashed black lines to indicate the region where the haplotype's LOD score is within a certain drop of its QTL peak. If this paramter is given values, two violet-red lines may be plotted. One violet-red line is the SNP with the maximum LOD score within the QTL peak drop region and the second violet-red line is the SNP with the maximum LOD outside of the QTL peak drop region (only if the SNP inside the region does not have the overall maximum LOD score). Multiple lines may be plotted if multiple SNPs have the maximum LOD score. A second optional parameter, haplo_scan, allows plotting of the \code{\link[=scan1]{scan1()}} output around the genomic interval after the scan1 object is filtered with \code{\link[=scan1_snpasso_rng]{scan1_snpasso_rng()}}.
}


\arguments{
\item{scan1output}{Output of \code{\link[=scan1]{scan1()}} using
SNP probabilities derived by
\code{\link[=genoprob_to_snpprob]{genoprob_to_snpprob()}}.}

\item{snpinfo}{Data frame with SNP information with the following
columns (the last three are generally derived from with
\code{\link[=index_snps]{index_snps()}}):
\itemize{
\item \code{chr} - Character string or factor with chromosome
\item \code{pos} - Position (in same units as in the \code{"map"}
attribute in \code{genoprobs}.
\item \code{sdp} - Strain distribution pattern: an integer, between
1 and \eqn{2^n - 2} where \eqn{n} is the number of strains, whose
binary encoding indicates the founder genotypes
\item \code{snp} - Character string with SNP identifier (if
missing, the rownames are used).
\item \code{index} - Indices that indicate equivalent
groups of SNPs.
\item \code{intervals} - Indexes that indicate which marker
intervals the SNPs reside.
\item \code{on_map} - Indicate whether SNP coincides with a marker
in the \code{genoprobs}
}}


\item{max_peak}{Numeric value in Mbps of the location of the haplotype's QTL peak. This can be obtained from \code{\link[=max_scan1]{max_scan1()}} 'pos' column.}

\item{haplo_peak_drop}{A vector of length two containing the range of Mbps that have LOD scores within the QTL peak drop. This vector should be the output from \code{\link[=lod_drop_rng]{lod_drop_rng()}}.}

\item{haplo_scan}{A dataframe containing the haplotype's LOD score and position around the genomic interval. The first column should contain the haplotype's LOD score, while the second column should be titled 'pos' and contain the Mbps positions. This dataframe should be the output from \code{\link[=scan1_snpasso_rng]{scan1_snpasso_rng()}}.}

\item{genes}{Optional data frame containing gene information for
the region, with columns \code{start} and \code{stop} in Mbp, \code{strand}
(as \code{"-"}, \code{"+"}, or \code{NA}), and \code{Name}. If included, a
two-panel plot is produced, with SNP associations above and
gene locations below.}

\item{lodcolumn}{LOD score column to plot (a numeric index, or a
character string for a column name). Only one value allowed.}

\item{show_all_snps}{If TRUE, expand to show all SNPs.}

\item{chr}{Vector of character strings with chromosome IDs to plot.}

\item{add}{If TRUE, add to current plot (must have same map and
chromosomes).}

\item{drop_hilit}{SNPs with LOD score within this amount of the maximum SNP association will be highlighted.}

\item{col_hilit}{Color of highlighted points}

\item{col}{Color of other points}

\item{gap}{Gap between chromosomes.}

\item{minlod}{Minimum LOD to display. (Mostly for GWAS, in which
case using \code{minlod=1} will greatly increase the plotting speed,
since the vast majority of points would be omittted.}

\item{...}{Additional graphics parameters.}
}
