% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snpasso_and_genes2.R
\name{plot_snpasso_and_genes2}
\alias{plot_snpasso_and_genes2}
\title{Two-panel plot with both SNP association and genes}
\usage{
plot_snpasso_and_genes2(scan1output, snpinfo, max_peak, haplo_peak_drop = NULL, haplo_scan = NULL, show_all_snps=TRUE,
                        drop_hilit=NA, col_hilit="violetred", col="darkslateblue",
                        gap=25, minlod=0,
                        genes, minrow=4, padding=0.2,
                        colors=c("black", "red3", "green4", "blue3", "darkorange2"),
                        scale_pos=1, start_field="start", stop_field="stop",
                        strand_field="strand", name_field="Name",
                        top_panel_prop=0.65, xlim=NULL, xaxt="s",
                        xlab=NULL, main="", sub="", ...)
}

\description{
Modified version of qtl2's internal plot_snpasso_and_genes function that is called by plot_snpasso2. See \code{\link[=plot_snpasso2]{plot_snpasso2}} and \code{\link[=plot_genes2]{plot_genes2}} for parameter info.
}


