% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan2_snpasso_rng.R
\name{scan2_snpasso_rng}
\alias{scan2_snpasso_rng}
\title{Find region with LOD scores around the given genomic interval}
\usage{
scan2_snpasso_rng(markers, scan1_output, max_chr, start, end)}

\description{
This function finds the haplotype region with LOD scores around the given genomic interval.
}


\arguments{
\item{markers}{A dataframe of marker positions. Should contain a 'chr' and 'pos' column.}
\item{scan1_output}{Output of \code{\link[=scan1]{scan1()}}.}
\item{max_chr}{Character value of the chromosome to be plotted.}
\item{start}{Position defining the start of an interval to get scan1 LOD scores.
Should be a single number, and if provided, \code{chr} should also
have length 1.}

\item{end}{Position defining the end of an interval to get scan1 LOD scores.
Should be a single number, and if provided, \code{chr} should also
have length 1.}
}
