% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan2blup.R
\name{scan2blup}
\alias{scan2blup}
\title{Calculate BLUPs of QTL effects in scan along one chromosome}
\usage{
scan2blup(chr, marker.id, genoprobs, pheno, kinship = NULL, addcovar = NULL,
    intcovar = NULL, samples = NULL, covar.factors = NULL, nullcovar = NULL, contrasts = NULL, se = FALSE, reml = TRUE,
    tol = 0.000000000001, cores = 1, quiet = TRUE)
}

\description{
Fixing this function....
}


