% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snpasso2.R
\name{plot_snpasso2}
\alias{plot_snpasso2}
\title{Plot SNP association#'}
\usage{
plot_snpasso2(scan1output, snpinfo, max_peak, haplo_peak_drop = NULL,
  haplo_scan = NULL, genes = NULL, lodcolumn = 1,
  show_all_snps = TRUE, chr = NULL, add = FALSE, drop_hilit = NA,
  col_hilit = "violetred", col = "darkslateblue", gap = 25,
  minlod = 0, ...)
}
\description{
Plot SNP association#'
}
